exercises: Update song exercises

+++ generators/Exercises/TwelveDays.cs
@@ -16,7 +16,8 @@ namespace Generators.Exercises
             {
                 canonicalDataCase.UseVariableForExpected = true;
                 canonicalDataCase.Expected = ConvertHelper.ToMultiLineString(canonicalDataCase.Expected);
-                if ((int)canonicalDataCase.Input["startVerse"] == (int)canonicalDataCase.Input["endVerse"])
+
+                if (canonicalDataCase.Input["startVerse"] == canonicalDataCase.Input["endVerse"])
                 {
                     canonicalDataCase.SetInputParameters("startVerse");
                 }
